1.新增角色，分组，岗位时保存对应的部门ID与数据域2.修复公共JS中删除后无法关闭弹层的BUG3.角色保存时增加UCODE否则提示非法提交4.修复规格页面显示BUG5.商品类型，商品属性，商品品牌，商品分类，商品供应商，商品规格，商品规格值删除记录时验证是否有子数据，有则提示无法删除

+++ ccloud-web-admin/src/main/java/org/ccloud/controller/admin/_BrandController.java
@@ -27,8 +27,10 @@ import org.ccloud.route.RouterMapping;
 import org.ccloud.route.RouterNotAllowConvert;
 import org.ccloud.utils.StringUtils;
 import org.ccloud.model.Brand;
+import org.ccloud.model.GoodsCategory;
 import org.ccloud.model.Supplier;
 import org.ccloud.model.query.BrandQuery;
+import org.ccloud.model.query.GoodsCategoryQuery;
 import org.ccloud.model.query.SupplierQuery;
 
 import com.google.common.collect.ImmutableMap;
@@ -79,6 +81,11 @@ public class _BrandController extends JBaseCRUDController<Brand> {
 	@RequiresPermissions(value={"/admin/brand/edit","/admin/all"},logical=Logical.OR)
 	public void delete() {
 		String id = getPara("id");
+		List<GoodsCategory> list = GoodsCategoryQuery.me().findCategoryByBrandId(id);
+		if (list.size() > 0) {
+			renderAjaxResultForError("已有分类在此品牌下");
+			return;
+		}
 		final Brand r = BrandQuery.me().findById(id);
 		if (r != null) {
 			if (r.delete()) {
